+++ Src/Library/changelog.md
@@ -32,6 +32,7 @@ app.UseFastEndpoints(c =>
 
 ### NEW
 - code snippets added to [visual studio extension](https://marketplace.visualstudio.com/items?itemName=dj-nitehawk.FastEndpoints)
+- `FastEndpoints.ClientGen` package for c# and typescript client generation with `NSwag`
 - override default model binding logic by inheriting from `RequestBinder<TRequest>` class #189
 - `DontCatchExceptions()` method to enable custom exception handler middleware #186
 - `Validator<TValidafor>()` method for being explicit in the endpoint configuration #190
@@ -41,7 +42,6 @@ app.UseFastEndpoints(c =>
 - `ProducesProblemFE()` extension method for `RouteHandlerBuilder` [#info](https://discord.com/channels/933662816458645504/1004762111546769498)
 - ability to customize permissions claim type #187
 - multiple route support for http attributes #129
-- `FastEndpoints.ClientGen` package for c# and typescript client generation with `NSwag`
 
 ### IMPROVEMENTS
 - remove the `new()` contraint on response dtos so a parameterless ctor is not needed on response classes #184